+++ src/main/java/racing/persistence/h2/game/H2RacingGameDao.java
@@ -1,4 +1,4 @@
-package racing.persist.game;
+package racing.persistence.h2.game;
 
 import java.sql.PreparedStatement;
 import java.sql.Statement;
